[FEAT] : 본 서버 연동 및 수정/추가/조회 구현

+++ domain/src/commonMain/kotlin/bff/wespot/staff/domain/vote/VoteRepository.kt
@@ -3,5 +3,5 @@ package bff.wespot.staff.domain.vote
 interface VoteRepository {
     suspend fun getVoteQuestions(): Result<List<VoteQuestion>>
     suspend fun postVoteQuestion(question: VoteQuestionContent): Result<Unit>
-    suspend fun editVoteQuestion(id: Int, question: VoteQuestionContent): Result<Unit>
+    suspend fun editVoteQuestion(id: Long, question: VoteQuestionContent): Result<Unit>
 }
